Set QT as backend to matplotlib earlier
When importing matplotlib for the first time, set backend to
QT4.  If plot is imported or called prior to matplotlib.use(),
the .use() call will be useless.
